add manager endpoints and logic

+++ server/services/projects.py
@@ -194,5 +194,73 @@ def remove_member_by_id(sender_id: str, project_id: str, member_id: str):
     if member_id not in project.members:
         raise InvalidParameter("User is not a member of the project")
 
+    if member_id in project.managers:
+        result = project_dao.unassign_manager_from_project(project_id, member_id)
+    
+    if result.modified_count == 0:
+        raise InvalidParameter("Error removing user as manager")
+
     project_dao.remove_member_from_project(project_id, member_id)
+    return True
+
+def assign_manager_role_to_user_by_id(sender_id: str, project_id: str, member_id: str):
+    """
+    Assigns a user as a manager of a project.
+
+    :param sender_id: The unique identifier of the user assigning the manager role.
+    :type sender_id: str
+
+    :param project_id: The unique identifier of the project.
+    :type project_id: str
+
+    :param member_id: The unique identifier of the user being assigned as a manager.
+    :type member_id: str
+
+    :raises ProjectNotFound: If no project is found with the provided ID.
+    """
+    if not project_dao.is_project_exist(project_id):
+        raise ProjectNotFound(project_id)
+
+    project = project_dao.get_project(project_id)
+    if project.owner != sender_id:
+        raise InvalidParameter("Only the project owner can assign managers")
+
+    if member_id not in project.members:
+        raise InvalidParameter("User is not a member of the project")
+
+    if member_id in project.managers:
+        raise InvalidParameter("User is already a manager of the project")
+
+    project_dao.assign_manager_to_project(project_id, member_id)
+    return True
+
+def unassign_member_role_from_user_by_id(sender_id: str, project_id: str, member_id: str):
+    """
+    Unassigns a user from being a manager of a project.
+
+    :param sender_id: The unique identifier of the user unassigning the manager role.
+    :type sender_id: str
+
+    :param project_id: The unique identifier of the project.
+    :type project_id: str
+
+    :param member_id: The unique identifier of the user being unassigned as a manager.
+    :type member_id: str
+
+    :raises ProjectNotFound: If no project is found with the provided ID.
+    """
+    if not project_dao.is_project_exist(project_id):
+        raise ProjectNotFound(project_id)
+
+    project = project_dao.get_project(project_id)
+    if project.owner != sender_id:
+        raise InvalidParameter("Only the project owner can unassign managers")
+
+    if member_id not in project.members:
+        raise InvalidParameter("User is not a member of the project")
+
+    if member_id not in project.managers:
+        raise InvalidParameter("User is not a manager of the project")
+
+    project_dao.unassign_manager_from_project(project_id, member_id)
     return True
